Rename GitHubApiRepository with proper casing
